Fix issues in CommonEntity
Fix version increments.
Refactoring EntityListeners implements.
