--- src/main/java/jp/mydns/projectk/safi/entity/FooterContext.java
@@ -1,62 +1,0 @@
-/*
- * Copyright (c) 2025, Project-K
- * All rights reserved.
- *
- * Redistribution and use in source and binary forms, with or without
- * modification, are permitted provided that the following conditions are met:
- *
- * * Redistributions of source code must retain the above copyright notice, this
- *   list of conditions and the following disclaimer.
- * * Redistributions in binary form must reproduce the above copyright notice,
- *   this list of conditions and the following disclaimer in the documentation
- *   and/or other materials provided with the distribution.
- *
- * THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS "AS IS"
- * AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE
- * IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE
- * ARE DISCLAIMED. IN NO EVENT SHALL THE COPYRIGHT HOLDER OR CONTRIBUTORS BE
- * LIABLE FOR ANY DIRECT, INDIRECT, INCIDENTAL, SPECIAL, EXEMPLARY, OR
- * CONSEQUENTIAL DAMAGES (INCLUDING, BUT NOT LIMITED TO, PROCUREMENT OF
- * SUBSTITUTE GOODS OR SERVICES; LOSS OF USE, DATA, OR PROFITS; OR BUSINESS
- * INTERRUPTION) HOWEVER CAUSED AND ON ANY THEORY OF LIABILITY, WHETHER IN
- * CONTRACT, STRICT LIABILITY, OR TORT (INCLUDING NEGLIGENCE OR OTHERWISE)
- * ARISING IN ANY WAY OUT OF THE USE OF THIS SOFTWARE, EVEN IF ADVISED OF THE
- * POSSIBILITY OF SUCH DAMAGE.
- */
-package jp.mydns.projectk.safi.entity;
-
-import java.time.LocalDateTime;
-
-/**
- * Provides a footer values for {@link CommonEntity}.
- *
- * @author riru
- * @version 3.0.0
- * @since 3.0.0
- */
-public interface FooterContext {
-
-    /**
-     * Get real time of UTC time zone. Rounded down to the nearest millisecond.
-     *
-     * @return real time
-     * @since 3.0.0
-     */
-    LocalDateTime getUtcNow();
-
-    /**
-     * Get login account id.
-     *
-     * @return account id
-     * @since 3.0.0
-     */
-    String getAccountId();
-
-    /**
-     * Get current process name.
-     *
-     * @return process name
-     * @since 3.0.0
-     */
-    String getProcessName();
-}
